vedanta-main/nephrology_agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
streamlit==1.40.2
python-dotenv==1.1.1
requests==2.32.5
python-multipart==0.0.20